requirements.txt
asgiref==3.5.2
boto3==1.24.70
botocore==1.27.70
cachetools==5.2.0
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
cryptography==38.0.1
decorator==4.4.2
Django==4.1.1
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.13.1
docutils==0.19
environ==1.0
errata-tool==1.28.1
idna==2.10
jmespath==1.0.1
jsonpath-rw==1.4.0
kerberos==1.3.1
mysqlclient==2.1.1
ply==3.11
pycparser==2.21
pykerberos==1.2.4
python-dateutil==2.8.1
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
requests==2.24.0
requests-kerberos==0.14.0
s3transfer==0.6.0
six==1.15.0
sqlparse==0.4.2
urllib3==1.25.8
schedule==1.1.0
django-filter==22.1